+++ src/native/os_manager_native.c
@@ -1,7 +1,7 @@
 #include "commands.h"
 #include "os_manager.h"
-#include <arpa/inet.h>
-#include <netinet/in.h>
+#include "socket.h"
+#include "string.h"
 #include <pthread.h>
 #include <sched.h>
 #include <stdio.h>
@@ -9,19 +9,17 @@
 #include <time.h>
 #include <unistd.h>
 
-static int sockfd;
-
 Mutex_t create_mutex()
 {
     pthread_mutex_t* mutex = (pthread_mutex_t*)malloc(sizeof(pthread_mutex_t));
     if (mutex == NULL) {
         perror("mutex malloc failed");
-        return NULL;
+        exit(1);
     }
     if (pthread_mutex_init(mutex, NULL) != 0) {
         perror("pthread_mutex_init failed");
         free(mutex);
-        return NULL;
+        exit(1);
     }
     return (Mutex_t)mutex;
 }
@@ -46,14 +44,12 @@ Thread_t create_thread(ThreadFunc_t func, void* argument)
     pthread_t* new_thread = malloc(sizeof(pthread_t));
     if (new_thread == NULL) {
         perror("thread malloc failed");
-        while (1)
-            ;
+        exit(1);
     }
     if (pthread_create(new_thread, NULL, func, argument) != 0) {
         perror("thread creation failed");
         free(new_thread);
-        while (1)
-            ;
+        exit(1);
     }
     return (Thread_t)new_thread;
 }
@@ -61,91 +57,6 @@ Thread_t create_thread(ThreadFunc_t func, void* argument)
 uint32_t get_current_timeMs()
 {
     return (clock() * 1000) / CLOCKS_PER_SEC;
-}
-
-void init_address(struct sockaddr_in* serv_addr, char* addr, uint16_t port)
-{
-    serv_addr->sin_family = AF_INET;
-    serv_addr->sin_addr.s_addr = inet_addr(addr);
-    serv_addr->sin_port = htons(port);
-}
-
-void init_socket(struct sockaddr_in serv_addr, uint16_t port)
-{
-    sockfd = socket(AF_INET, SOCK_STREAM, 0);
-    if (sockfd < 0) {
-        printf("Unable to create socket\n");
-        exit(1);
-    }
-    if (connect(sockfd, (struct sockaddr*)&serv_addr, sizeof(serv_addr)) < 0) {
-        printf("Unable to connect to the server\n");
-        exit(1);
-    }
-    printf("Connected to the server successfully\n");
-}
-
-uint16_t len(const char* str)
-{
-    uint16_t length = 0;
-    while (*str != '\0') {
-        length++;
-        str++;
-    }
-    return length;
-}
-
-void str_copy(const char* src, char* dest)
-{
-    while (*src != '\0') {
-        *dest = *src;
-        dest++;
-        src++;
-    }
-    *dest = '\0';
-}
-
-void add_endline(const char* str, char* str_result)
-{
-    while (*str != '\0') {
-        *str_result = *str;
-        str_result++;
-        str++;
-    }
-    *str_result = '\n';
-    str_result++;
-    *str_result = '\0';
-}
-
-void send_message(const char* message)
-{
-    char message_with_endline[MAX_COMMAND_SIZE];
-    if (message[len(message) - 1] != '\n') {
-        add_endline(message, message_with_endline);
-    } else {
-        str_copy(message, message_with_endline);
-    }
-    printf("sending: %s\n", message_with_endline);
-    if (send(sockfd, message_with_endline, len(message_with_endline), 0) < 0) {
-        printf("Unable to send message\n");
-        exit(1);
-    }
-}
-
-void receive_message(char* response)
-{
-    if (recv(sockfd, response, MAX_RESPONSE_SIZE, 0) < 0) {
-        printf("Unable to receive message\n");
-        exit(1);
-    }
-    char* original_response = response;
-    char c = -1;
-    while (c != '\n') {
-        c = *response;
-        response++;
-    }
-    *response = '\0';
-    response = original_response;
-    printf("received: %s\n", response);
 }
 
 int puts(const char* str)
@@ -184,8 +95,8 @@ void os_initialisation(int argc, char* argv[])
         char* localhost = "127.0.0.1";
         str_copy(localhost, address);
     }
-    if (inet_addr(address) == INADDR_NONE) {
-        printf("Invalid address\n");
+    if (!is_address_valid(address)) {
+        perror("Invalid address\n");
         exit(1);
     }
     uint16_t port = (uint16_t)atoi(argv[2]);
@@ -199,7 +110,7 @@ void os_initialisation(int argc, char* argv[])
     char response[MAX_RESPONSE_SIZE] = { 0 };
     receive_message(response);
     if (!(response[0] == 'O' && response[1] == 'K' && response[2] == '\n' && response[3] == '\0')) {
-        printf("Connection refused by the server");
+        perror("Connection refused by the server");
         exit(1);
     }
 }
